requirements.txt
fastapi
sqlalchemy
httpx
uvicorn
pyyaml
ruff
